moved dleq into independant subpackage
